homebridge integration and noble
